+++ tests/testcases_output/php/ok_type_helpers/docs.php
@@ -1,0 +1,12 @@
+<?php
+global $pasdoc;
+$pasdoc = array(
+  'ok_type_helpers' => array('html_filename' => 'ok_type_helpers.html', 'type' => 'unit'),
+  'TMyEnum' => array('html_filename' => 'ok_type_helpers.html#TMyEnum', 'type' => 'type'),
+  'one' => array('html_filename' => 'ok_type_helpers.html#one', 'type' => 'enum constant'),
+  'two' => array('html_filename' => 'ok_type_helpers.html#two', 'type' => 'enum constant'),
+  'three' => array('html_filename' => 'ok_type_helpers.html#three', 'type' => 'enum constant'),
+  'TStringHelper' => array('html_filename' => 'ok_type_helpers.TStringHelper.html', 'type' => 'record'),
+  'TMyEnumHelper' => array('html_filename' => 'ok_type_helpers.TMyEnumHelper.html', 'type' => 'record'),
+  'TMyEnumHelper.GetNext' => array('html_filename' => 'ok_type_helpers.TMyEnumHelper.html#GetNext', 'type' => 'function'),
+);